\name{RCourseLab2-package}
\alias{RCourseLab2-package}
\alias{RCourseLab2}
\docType{package}
\title{
\packageTitle{RCourseLab2}
}
\description{
\packageDescription{RCourseLab2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCourseLab2}
\packageIndices{RCourseLab2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RCourseLab2}

Maintainer: \packageMaintainer{RCourseLab2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
